# docs/swagger.yaml
basePath: /
definitions:
  Server.Subscription:
    properties:
      end_date:
        type: string
      id:
        type: string
      price:
        type: integer
      service_name:
        type: string
      start_date:
        example: "2025-09-01T00:00:00Z"
        type: string
      user_id:
        type: string
    type: object
  Server.SubscriptionRequest:
    properties:
      price:
        type: integer
      service_name:
        type: string
      start_date:
        type: string
      user_id:
        type: string
    type: object
  Server.ToFilter:
    properties:
      service_name:
        type: string
      user_id:
        type: string
    type: object
host: localhost:8080
info:
  contact: {}
  description: This is a sample API for managing subscriptions.
  title: Subscription API
  version: "1.0"
paths:
  /subscriptions:
    get:
      description: Retrieve a list of all subscriptions
      operationId: get-all-subscriptions
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/Server.Subscription'
            type: array
      summary: Get all subscriptions
      tags:
      - subscriptions
    post:
      consumes:
      - application/json
      description: Add a new subscription to the database
      operationId: create-subscription
      parameters:
      - description: Subscription data
        in: body
        name: subscription
        required: true
        schema:
          $ref: '#/definitions/Server.SubscriptionRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/Server.Subscription'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Create a new subscription
      tags:
      - subscriptions
  /subscriptions/{id}:
    delete:
      consumes:
      - application/json
      description: Remove a subscription by ID
      operationId: delete-subscription
      parameters:
      - description: Subscription ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Delete a subscription
      tags:
      - subscriptions
    patch:
      consumes:
      - application/json
      description: Update subscription details by ID
      operationId: update-subscription
      parameters:
      - description: Subscription ID
        in: path
        name: id
        required: true
        type: integer
      - description: Updated subscription data
        in: body
        name: subscription
        required: true
        schema:
          $ref: '#/definitions/Server.SubscriptionRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/Server.Subscription'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Update an existing subscription
      tags:
      - subscriptions
  /subscriptionsByFilter:
    get:
      consumes:
      - application/json
      description: Retrieve subscriptions based on certain filters
      operationId: get-subscriptions-by-filter
      parameters:
      - description: Filter criteria
        in: body
        name: filter
        required: true
        schema:
          $ref: '#/definitions/Server.ToFilter'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/Server.Subscription'
            type: array
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get subscriptions by filter
      tags:
      - subscriptions
swagger: "2.0"